starting to do a rendering test

+++ OSMTrackReader/src/osmtrackreader/OSMTrackReader.java
@@ -26,30 +26,30 @@ public class OSMTrackReader {
      */
     public static void main(String[] args) {
         // TODO code application logic here
-        
+
         final List points = GpxFileDataAccess.getPoints(new File("./track.gpx"));
-        
-        final Trip trip = Trip.makeTrip(1, 
+
+        final Trip trip = Trip.makeTrip(1,
                 new TrackSegment(points, TrackSegment.caminarType.undef));
-        
+
         final List allSegments = trip.getSegments();
-        
+
         final List<TrackPoint> allPoints = trip.getPoints();
-        
+
         ListIterator<TrackPoint> i;
         i = allPoints.listIterator();
-            
-        
+
+
         // TODO: Path erstellen und dann folgende Vektoren in den Pfad aufnehmen
-        
+
         while (i.hasNext()) {
             TrackPoint p = (TrackPoint) i.next();
             System.out.printf("%f, %f, %d \n", p.getLat(), p.getLon(), p.getElevation());
             Vector3 vec = new Vector3(
-                    (float) p.getLat(), 
-                    (float) p.getLon(), 
-                    (float) (p.getElevation() / 100.0));
-            
+                    (float) p.getLat(),
+                    (float) p.getLon(),
+                    p.getElevation() / 100f);
+
         }
     }
 }
